requirements.txt
elasticsearch==6.3.1
elasticsearch-curator==5.5.4
voluptuous>=0.9.3
boto3>=1.7.24
requests_aws4auth>=0.9
voluptuous>=0.9.3
click>=6.7
certifi>=2018.4.16
six>=1.11.0
PySocks==1.6.8
urllib3==1.23
